Рефакторинг

+++ AnimateLab/rects.cpp
@@ -106,9 +106,9 @@ void transparencyChenge(Rects &rects, const float &elapsedTime, int mode) // mod
 		rect = &rects.items[itemNumber];
 
 		if (mode == 0)
-			newAlpha = RECT_START_COLOR.a - static_cast<sf::Uint8>((int)rects.alphaChenge);
+			newAlpha = RECT_START_COLOR.a - static_cast<sf::Uint8>(rects.alphaChenge);
 		if (mode == 1)
-			newAlpha = RECT_START_COLOR.a - static_cast<sf::Uint8>(TRANSPARENCY_CHENGE) + static_cast<sf::Uint8>((int)rects.alphaChenge);
+			newAlpha = RECT_START_COLOR.a - static_cast<sf::Uint8>(TRANSPARENCY_CHENGE) + static_cast<sf::Uint8>(rects.alphaChenge);
 
 		rect->setFillColor(sf::Color(RECT_START_COLOR.r, RECT_START_COLOR.g, RECT_START_COLOR.b, newAlpha));
 	}
